Split source tests from 1 to 3 files.
